environment: Move GetStdin to environment
