+++ internal/cli/cmds/provisioning/server_system_network.go
@@ -2,21 +2,19 @@ package provisioning
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"io"
 	"os"
-	"os/exec"
-	"strings"
 
 	incusosapi "github.com/lxc/incus-os/incus-osd/api"
-	"github.com/lxc/incus/v6/shared/revert"
 	"github.com/lxc/incus/v6/shared/termios"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v3"
 
 	"github.com/FuturFusion/operations-center/internal/cli/validate"
 	"github.com/FuturFusion/operations-center/internal/client"
+	"github.com/FuturFusion/operations-center/internal/editor"
+	"github.com/FuturFusion/operations-center/internal/environment"
 	"github.com/FuturFusion/operations-center/shared/api"
 )
 
@@ -101,7 +99,7 @@ func (c *cmdServerSystemNetworkEdit) Run(cmd *cobra.Command, args []string) erro
 	name := args[0]
 
 	// If stdin isn't a terminal, read text from it.
-	if !termios.IsTerminal(getStdinFd()) {
+	if !termios.IsTerminal(environment.GetStdinFd()) {
 		contents, err := io.ReadAll(os.Stdin)
 		if err != nil {
 			return err
@@ -135,7 +133,7 @@ func (c *cmdServerSystemNetworkEdit) Run(cmd *cobra.Command, args []string) erro
 	}
 
 	// Spawn the editor
-	content, err := textEditor("", append([]byte(c.helpTemplate()+"\n\n"), b.Bytes()...))
+	content, err := editor.Spawn("", append([]byte(c.helpTemplate()+"\n\n"), b.Bytes()...))
 	if err != nil {
 		return err
 	}
@@ -159,7 +157,7 @@ func (c *cmdServerSystemNetworkEdit) Run(cmd *cobra.Command, args []string) erro
 				return err
 			}
 
-			content, err = textEditor("", content)
+			content, err = editor.Spawn("", content)
 			if err != nil {
 				return err
 			}
@@ -171,88 +169,4 @@ func (c *cmdServerSystemNetworkEdit) Run(cmd *cobra.Command, args []string) erro
 	}
 
 	return nil
-}
-
-// Spawn the editor with a temporary YAML file for editing configs.
-func textEditor(inPath string, inContent []byte) ([]byte, error) {
-	var f *os.File
-	var err error
-	var path string
-
-	// Detect the text editor to use
-	editor := os.Getenv("VISUAL")
-	if editor == "" {
-		editor = os.Getenv("EDITOR")
-		if editor == "" {
-			for _, p := range []string{"editor", "vi", "emacs", "nano"} {
-				_, err := exec.LookPath(p)
-				if err == nil {
-					editor = p
-					break
-				}
-			}
-			if editor == "" {
-				return []byte{}, errors.New("No text editor found, please set the EDITOR environment variable")
-			}
-		}
-	}
-
-	if inPath == "" {
-		// If provided input, create a new file
-		f, err = os.CreateTemp("", "incus_editor_")
-		if err != nil {
-			return []byte{}, err
-		}
-
-		reverter := revert.New()
-		defer reverter.Fail()
-
-		reverter.Add(func() {
-			_ = f.Close()
-			_ = os.Remove(f.Name())
-		})
-
-		err = os.Chmod(f.Name(), 0o600)
-		if err != nil {
-			return []byte{}, err
-		}
-
-		_, err = f.Write(inContent)
-		if err != nil {
-			return []byte{}, err
-		}
-
-		err = f.Close()
-		if err != nil {
-			return []byte{}, err
-		}
-
-		path = fmt.Sprintf("%s.yaml", f.Name())
-		err = os.Rename(f.Name(), path)
-		if err != nil {
-			return []byte{}, err
-		}
-
-		reverter.Success()
-		reverter.Add(func() { _ = os.Remove(path) })
-	} else {
-		path = inPath
-	}
-
-	cmdParts := strings.Fields(editor)
-	cmd := exec.Command(cmdParts[0], append(cmdParts[1:], path)...)
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err = cmd.Run()
-	if err != nil {
-		return []byte{}, err
-	}
-
-	content, err := os.ReadFile(path)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	return content, nil
 }
